refactor(bot-client): address PR #856 review — drop cast + re-exports
- NormalizationStep.ts: remove redundant \`(ref as { timestamp: string })\`
  cast. The surrounding type already allows writing to \`timestamp\`, so
  the cast was pure noise (matching the msg.createdAt cleanup earlier
  in the same function).

- Drop the CharacterData re-export from commands/character/config.ts:
  update all 15 internal callers to import from \`./characterTypes.js\`
  directly. config.ts keeps the type import for its own use.

- Drop the DenylistEntryResponse re-export from commands/deny/browse.ts:
  update the one external caller (view.ts) to import from
  \`./browseTypes.js\` directly.

Both re-exports were only load-bearing because callers hadn't been
migrated yet. Cleaning them up avoids the dual-import pattern the
reviewer flagged and keeps the leaf modules as the single source of
truth for the types.
