+++ services/bot-client/src/commands/character/dashboardActions.ts
@@ -21,9 +21,9 @@ import {
 import {
   getCharacterDashboardConfig,
   buildCharacterDashboardOptions,
-  type CharacterData,
   type CharacterSessionData,
 } from './config.js';
+import type { CharacterData } from './characterTypes.js';
 import { toGatewayUser } from '../../utils/userGatewayClient.js';
 import { fetchCharacter, updateCharacter, toggleVisibility } from './api.js';
